fix virtual table glitches
