+++ src/components/table/Td.tsx
@@ -4,7 +4,7 @@ import {
   tableFillLevelToBorder,
   tableFillLevelToHighlightedCellBg,
 } from './colors'
-import { type TableFillLevel } from './Table'
+import { type TableFillLevel } from './tableUtils'
 
 export const Td = styled.td<{
   $fillLevel: TableFillLevel
